refactor: 필요없는 story 파일 삭제 및 stories 파일 위치 이동
