+++ src/components/CardBox.stories.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { ComponentMeta, ComponentStory } from '@storybook/react';
 
-import { CardBox } from '../../components';
+import { CardBox } from './index';
 
 export default {
   title: 'Payments/CardBox',
@@ -9,10 +9,10 @@ export default {
   parameters: {
     // More on Story layout: https://storybook.js.org/docs/react/configure/story-layout
     layout: 'centered',
-  },
+  }
 } as ComponentMeta<typeof CardBox>;
-
 const Template: ComponentStory<typeof CardBox> = (args) => <CardBox {...args} />;
+
 const defaultData = {
   cardCompany: '클린카드',
   cardNumber: '1111 - 2222 - oooo - oooo',
@@ -28,7 +28,6 @@ Empty.args = {
 export const Small = Template.bind({});
 Small.args = {
   ...defaultData,
-  type: 'small',
 };
 
 export const Big = Template.bind({});
