+++ src/components/index.ts
@@ -1,3 +1,3 @@
-export { default as Input } from './Input';
+export { default as Input } from './form/Input';
 export { default as CardBox } from './CardBox';
 export { default as Modal } from './Modal';
