--- src/stories/Color.stories.mdx
@@ -1,35 +1,0 @@
-import { ColorItem, ColorPalette, Meta } from '@storybook/addon-docs';
-
-<Meta title="Colors"/>
-
-<ColorPalette>
-  <ColorItem
-    title="theme.color.greyscale"
-    subtitle="Some of the greys"
-    colors={{ White: '#FFFFFF', Alabaster: '#F8F8F8', Concrete: '#F3F3F3' }}
-  />
-  <ColorItem
-    title="theme.color.primary"
-    subtitle="Coral"
-    colors={{ WildWatermelon: '#FF4785' }}
-  />
-  <ColorItem
-    title="theme.color.secondary"
-    subtitle="Ocean"
-    colors={{ DodgerBlue: '#1EA7FD' }}
-  />
-  <ColorItem
-    title="theme.color.card"
-    subtitle="Card Color"
-    colors={{
-      Card1: '#ef5350',
-      Card2: '#ab47bc',
-      Card3: '#5c6bc0',
-      Card4: '#29b6f6',
-      Card5: '#26a69a',
-      Card6: '#9ccc65',
-      Card7: '#ffee58',
-      Card8: '#ffa726',
-    }}
-  />
-</ColorPalette>
